@@ -1,6 +1,8 @@
 package com.blog.model;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
+
 
 @Entity
 public class Blog {
